Switch to 2D array
This makes a lot of the grid access simpler, at the cost of making
initial seeding harder. The pay-off seems worth it, as the simpler logic
will likely reduce bugs (in fact, I think the symmetry bug that I was
seeing before is gone now; although there is a new bug in its place,
which shows the initial state of the grid being initialized not quite as
randomly as I would hope).

Signed-off-by: Wincent Colaiuta <win@wincent.com>